% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaCorrelationTable.R
\name{apa.cor.table}
\alias{apa.cor.table}
\title{Creates a correlation table in APA style with means and standard deviations}
\usage{
apa.cor.table(
  data,
  filename = NA,
  table.number = NA,
  show.conf.interval = TRUE,
  show.sig.stars = TRUE,
  show.pvalue = TRUE,
  landscape = TRUE
)
}
\arguments{
\item{data}{Project data frame}

\item{filename}{(optional) Output filename document filename (must end in .rtf or .doc only)}

\item{table.number}{Integer to use in table number output line}

\item{show.conf.interval}{(TRUE/FALSE) Display confidence intervals in table. This argument is deprecated and will be removed from later versions.}

\item{show.sig.stars}{(TRUE/FALSE) Display stars for significance in table.}

\item{show.pvalue}{(TRUE/FALSE) Display p-value in table.}

\item{landscape}{(TRUE/FALSE) Make RTF file landscape}
}
\value{
APA table object
}
\description{
Creates a correlation table in APA style with means and standard deviations
}
\examples{
# View top few rows of attitude data set
head(attitude)

# Use apa.cor.table function
table1 <- apa.cor.table(attitude)


# Save Table 1 in a .doc document
apaTables:::apa.save(filename = "table1.doc", table1)


# Create a table for your PDF
# Include the lines below in your rmarkdown or Quarto document
apaTables:::apa.knit.table.for.pdf(table1)
# For better results:
# BEFORE the code chunk adjust the line spacing to 0.5
# \renewcommand{\arraystretch}{.5}
# AFTER the code chunk adjust the line spacing back to 1.0
# \renewcommand{\arraystretch}{1}

# delete demo file
if (file.exists("table1.doc")) {
     file.remove("table1.doc")
}
}
