% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfOutput.R
\name{apa.knit.table.for.pdf}
\alias{apa.knit.table.for.pdf}
\title{Create latex output for an apaTables object}
\usage{
apa.knit.table.for.pdf(
  table_object,
  table_note = NULL,
  table_title = NULL,
  line_spacing = 1
)
}
\arguments{
\item{table_object}{Previously constructed apaTable object}

\item{table_note}{Replace default table note with this text}

\item{table_title}{Replace default table title with this text}

\item{line_spacing}{Line spacing multiplier for table}
}
\value{
A kableExtra LaTeX table object for rendering in PDF documents.
}
\description{
Create latex output for an apaTables object
}
