% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfOutput.R
\name{apa.save}
\alias{apa.save}
\title{Save previously constructed APA table objects in a single .doc file}
\usage{
apa.save(filename, ..., paper = "us")
}
\arguments{
\item{filename}{Filename (e.g., my.tables.doc)}

\item{...}{apaTable objects to be saved}

\item{paper}{Use "us" or "a4". Default is "us".}
}
\value{
Save status
}
\description{
Save previously constructed APA table objects in a single .doc file
}
\examples{
library(apaTables)

table1 <- apa.1way.table(iv = dose, dv = libido,
               data = viagra, table.number = 1)

table2 <- apa.2way.table(iv1 = gender, iv2 = alcohol,
                         dv = attractiveness,
                         data = goggles, table.number = 1)

apa.save(filename = "my.tables.doc", table1, table2)

# delete demo file
if (file.exists("my.tables.doc")) {
     file.remove("my.tables.doc")
}
}
