% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dating_wide.R
\docType{data}
\name{dating_wide}
\alias{dating_wide}
\title{dating data from textbook}
\format{A data frame with 20 rows and 11 columns. Gender is a between subjects variable. Looks and Personality are within subject variables. Both gender and particpant are factors.
\describe{
  \item{participant}{Factor: Participant ID number}
  \item{gender}{Factor: Gender: Male/Female}
  \item{attractive_high}{Date rating where looks=attractive and personality=high}
  \item{average_high}{Date rating where looks=average and personality=high}
  \item{ugly_high}{Date rating where looks=ugly and personality=high}
  \item{attractive_some}{Date rating where looks=attractive and personality=some}
  \item{average_some}{Date rating where looks=average and personality=some}
  \item{ugly_some}{Date rating where looks=ugly and personality=some}
  \item{attractive_none}{Date rating where looks=attractive and personality=none}
  \item{average_none}{Date rating where looks=average and personality=none}
  \item{ugly_none}{Date rating where looks=ugly and personality=none}
}}
\source{
\url{http://studysites.sagepub.com/dsur/study/}
}
\usage{
data(dating_wide)
}
\description{
A data set from Field et al (2012)
}
\references{
Field, A., Miles, J., & Field, Z. (2012) Discovering Statistics Using R. Sage: Chicago.
}
\keyword{datasets}
