% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-drink_attitude_wide.R
\docType{data}
\name{drink_attitude_wide}
\alias{drink_attitude_wide}
\title{drink attitude data from textbook}
\format{A data frame with 20 rows and 10 columns. Drink and Imagery are within subject variables. Particpant is a factor.
\describe{
  \item{participant}{Factor: Participant ID number}
  \item{beer_positive}{Attitude where drink=beer and imagery=positive}
  \item{beer_negative}{Attitude where drink=beer and imagery=negative}
  \item{beer_neutral}{Attitude where drink=beer and imagery=neutral}
  \item{wine_positive}{Attitude where drink=wine and imagery=positive}
  \item{wine_negative}{Attitude where drink=wine and imagery=negative}
  \item{wine_neutral}{Attitude where drink=wine and imagery=neutral}
  \item{water_positive}{Attitude where drink=water and imagery=positive}
  \item{water_negative}{Attitude where drink=water and imagery=negative}
  \item{water_neutral}{Attitude where drink=water and imagery=neutral}
}}
\source{
\url{http://studysites.sagepub.com/dsur/study/}
}
\usage{
data(drink_attitude_wide)
}
\description{
A data set from Field et al (2012)
}
\references{
Field, A., Miles, J., & Field, Z. (2012) Discovering Statistics Using R. Sage: Chicago.
}
\keyword{datasets}
