% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-eysenck.R
\docType{data}
\name{Eysenck}
\alias{Eysenck}
\title{Eysenck data}
\format{A data frame with 100 rows and 3 variables:
\describe{
  \item{Age}{Young or Old}
  \item{Condition}{Experimental learning condition}
  \item{Recall}{Level of word recall}
}}
\source{
\url{http://www.uvm.edu/~dhowell/methods7/DataFiles/Tab13-2.dat}
}
\usage{
data(Eysenck)
}
\description{
Eysenck data
}
\references{
Howell, D. (2012). Statistical methods for psychology. Cengage Learning.
}
\keyword{datasets}

