% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fidler_thompson.R
\docType{data}
\name{fidler_thompson}
\alias{fidler_thompson}
\title{Fidler & Thompson (2001) Fixed-Effects ANOVA data}
\format{
A data frame with 24 rows and 3 variables:
\describe{
  \item{a}{Independent variable: a}
  \item{b}{Independent variable: b}
  \item{dv}{Dependent variable: dv}
}
}
\usage{
data(fidler_thompson)
}
\description{
A data set from Fidler & Thompson (2001)
}
\references{
Fidler, F. & Thompson, B. (2001). Computing correct confidence intervals for ANOVA fixed- and random-effects effect sizes. Educational and Psychological Measurement, 61, 575-604.
}
\keyword{datasets}
