% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-goggles.R
\docType{data}
\name{goggles}
\alias{goggles}
\title{goggles data from textbook}
\format{A data frame with 48 rows and 3 variables:
\describe{
  \item{gender}{Gender of participant}
  \item{alcohol}{Amount alcohol consumed }
  \item{attractiveness}{Perceived attractiveness}
}}
\source{
\url{http://studysites.sagepub.com/dsur/study/}
}
\usage{
data(goggles)
}
\description{
goggles data from textbook
}
\references{
Field, A., Miles, J., & Field, Z. (2012) Discovering Statistics Using R. Sage: Chicago.
}
\keyword{datasets}

