% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-viagra.R
\docType{data}
\name{viagra}
\alias{viagra}
\title{viagra data from textbook}
\format{A data frame with 15 rows and 2 variables:
\describe{
  \item{dose}{Level of viagra dose}
  \item{libido}{Libido after taking viagra}
}}
\source{
\url{http://studysites.sagepub.com/dsur/study/}
}
\usage{
data(viagra)
}
\description{
viagra data from textbook
}
\references{
Field, A., Miles, J., & Field, Z. (2012) Discovering Statistics Using R. Sage: Chicago.
}
\keyword{datasets}

