% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-album.R
\docType{data}
\name{album}
\alias{album}
\title{album data from textbook}
\format{A data frame with 200 rows and 4 variables:
\describe{
  \item{adverts}{Amount spent of adverts, thousands of pounds}
  \item{sales}{Album sales in thousands}
  \item{airplay}{Number of times songs from album played on radio week prior to release}
  \item{attract}{Attractiveness rating of band members}
}}
\source{
\url{http://studysites.sagepub.com/dsur/study/}
}
\usage{
data(album)
}
\description{
album data from textbook
}
\references{
Field, A., Miles, J., & Field, Z. (2012) Discovering Statistics Using R. Sage: Chicago.
}
\keyword{datasets}

