% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaANOVAFTable.R
\name{apa.anova.table}
\alias{apa.anova.table}
\title{Creates a ANOVA F table in APA style}
\usage{
apa.anova.table(lm_output, filename = NA, table.number = NA)
}
\arguments{
\item{filename}{(optional) Output filename document filename (must end in .rtf or .doc only)}

\item{table.number}{Integer to use in table number output line}

\item{...}{Regression (i.e., lm) result objects. Typically, one for each block in the regression.}

\item{is.random.predictors}{Indicate if predictors are random (TRUE/FALSE). Default is FALSE.}
}
\value{
APA table object
}
\description{
Creates a ANOVA F table in APA style
}
\examples{
# View top few rows of goggles data set
# from Discovering Statistics Using R
head(goggles)

#Use Helmert Contrasts as per SPSS
options(contrasts = c("contr.sum", "contr.poly"))

lm_output <- lm(attractiveness ~ gender*alcohol, data=goggles)
apa.anova.table(lm_output)

}

