% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaANOVAFTable.R
\name{apa.anova.table}
\alias{apa.anova.table}
\title{Creates a ANOVA F table in APA style}
\usage{
apa.anova.table(lm_output, filename, table.number = NA, conf.level = 0.9,
  type = 3)
}
\arguments{
\item{filename}{(optional) Output filename document filename (must end in .rtf or .doc only)}

\item{table.number}{Integer to use in table number output line}

\item{conf.level}{Level of confidence for interval around partial eta-squared (0 to 1). A value of .90 is the default, this helps to create consistency between the CI overlapping with zero and conclusions based on the p-value.}

\item{type}{Sum of Squares Type. Type II or Type III; specify, 2 or 3, respectively. Default value is 3.}

\item{...}{Regression (i.e., lm) result objects. Typically, one for each block in the regression.}
}
\value{
APA table object
}
\description{
Creates a ANOVA F table in APA style
}
\examples{
# Example from Fidler & Thompson (2001)
# You must set these contrasts to ensure values match SPSS
options(contrasts = c("contr.sum", "contr.poly"))
lm_output <- lm(dv ~ a*b, data=fidler_thompson)
apa.anova.table(lm_output)


}

